Add ability to track particles using envelope solver
The method uses the first two particles in the bunch to store the envelope parameters, and uses these parameters to apply space charge kicks to the rest of the particles in the bunch. this will likely be modified soon using Jeff's notes to include particles outside the beam envelope.

Also added monitor nodes which will write the envelope parameters to a file as well as the remaining bunch coordinates.